Licensing_API/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
python-multipart>=0.0.9
PyNaCl>=1.5.0
pydantic>=2.6.0
email-validator>=2.1.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

